pyproject.toml
[tool.poetry]
name = "ai-testtask"
version = "0.1.0"
description = "FastAPI ML"
authors = ["Sergey Leonovich <put_theport@mail.ru>"]
license = "MIT"
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"

# ML
torch = "^2.1.0"
torchvision = "^0.16.0"
scikit-learn = "^1.3.0"

# Обработка изображений
pillow = "^10.0.1"
opencv-python = "^4.8.1"

# Анализ данных
numpy = "^1.24.0"
pandas = "^2.0.0"
matplotlib = "^3.7.0"
seaborn = "^0.12.0"
jupyter = "^1.0.0"

# FastAPI
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.4.0"
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.6.0"
pre-commit = "^3.4.0"
jupyterlab = "^4.0.0"

[tool.poetry.scripts]
start = "app.main:main"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"